% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_mbb.R
\name{block_corrmat}
\alias{block_corrmat}
\title{Create a block correlation matrix}
\usage{
block_corrmat(rho = 0.8, blocks = 3, size = 2)
}
\arguments{
\item{rho}{numeric, correlation strength}

\item{blocks}{integer, number of blocks}

\item{size}{integer, block size}
}
\description{
Create a block correlation matrix
}
\details{
Implementation copied from
https://stats.stackexchange.com/questions/171342/more-efficient-way-to-create-block-toeplitz-matrix-in-r
}
