% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_mbb.R
\name{construct_mbeta}
\alias{construct_mbeta}
\title{Define a multivariate beta distribution}
\usage{
construct_mbeta(
  nu = 10,
  mu1 = 0.85,
  delta = 0.05,
  rho = 0.5,
  blocks = 1,
  size = 5,
  ...
)
}
\arguments{
\item{nu}{numeric, concentration parameter}

\item{mu1}{numeric, mean of first variable}

\item{delta}{numeric, distance between mean parameters}

\item{rho}{numeric, correlation parameter}

\item{blocks}{integer, specify block correlation structure}

\item{size}{integer, block size}

\item{...}{further arguments (ignored)}
}
\value{
A \code{SIMPle.dist} object with derived parameters,
compare \code{SIMPle::define_dist()}
}
\description{
Define a multivariate beta distribution
}
\details{
For details regarding the block correlation matrix, se \code{\link{block_corrmat}}
}
