% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_mle_core.R
\name{load_instance}
\alias{load_instance}
\title{Load MLE simulations instances from disk}
\usage{
load_instance(
  id = 1,
  filename = "mle_instance_",
  folder = "E:\\\\MLE_SIM\\\\DATA",
  sep = "\\\\",
  data = NULL,
  job = NULL
)
}
\arguments{
\item{id}{integer, id to append to \code{filename}}

\item{filename}{character, file name}

\item{folder}{character, file folder}

\item{sep}{character, path separator}

\item{data}{NULL (required for batchtools compatibility)}

\item{job}{NULL (required for batchtools compatibility)}
}
\description{
Load MLE simulations instances from disk
}
