% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_mle_core.R
\name{postproc_mle}
\alias{postproc_mle}
\title{Process MLE simulation results to investigate test decision characteristics}
\usage{
postproc_mle(
  results,
  analysis = c("acc", "cpe"),
  vals = seq(-0.05, 0.1, 0.01),
  delta = 0
)
}
\arguments{
\item{results}{data.frame with simulation results produced by \code{save_results}}

\item{analysis}{character, either "acc" or "cpe"}

\item{vals}{numeric, shift parameters to be investigated}

\item{delta}{numeric, difference between se0 and sp0 (default: 0, ignored for analysis="acc")}
}
\description{
Process MLE simulation results to investigate test decision characteristics
}
