% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_auc.R
\name{sample_auc}
\alias{sample_auc}
\title{Sample pseudo predictions from binormal AUC model}
\usage{
sample_auc(
  nV = 200,
  nE = 200,
  prev = 0.2,
  auc = 0.9,
  grd = 1000,
  data = NULL,
  job = NULL
)
}
\arguments{
\item{nV}{integer, validation sample size}

\item{nE}{integer, evaluation sample size}

\item{prev}{numeric, positive class prevalence}

\item{auc}{numeric, area under the curve}

\item{grd}{integer, size of grid}

\item{data}{ignored (required for batchtools compatibility)}

\item{job}{ignored (required for batchtools compatibility)}
}
\description{
Sample pseudo predictions from binormal AUC model
}
