% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_lfc.R
\name{sample_lfc_acc}
\alias{sample_lfc_acc}
\title{Generate artificial performance data (classification accuracy)}
\usage{
sample_lfc_acc(
  n = 100,
  S = 5,
  acc = 0.8,
  delta = 0,
  corr = "type=equi_rho=0",
  data = NULL,
  job = NULL
)
}
\arguments{
\item{n}{integer, sample size of evaluation study}

\item{S}{integer, number of classification models that are assessed}

\item{acc}{numeric, accuracy (between 0 and 1) of all models (under the global null), default
value is 0.8.}

\item{delta}{shift parameter}

\item{corr}{character string (default: "type=equi_rho=0"), specifies correlation structure,
see details}

\item{data}{ignored (required for batchtools compatability)}

\item{job}{ignored (required for batchtools compatability)}
}
\value{
a list with two entries: \code{args} (list of function arguments) and
\code{comp} (matrix of simulated correct (1) and false (0) predictions)
}
\description{
Generate binary data, simulating the assessment of classification accuracy.
Result may be passed to \code{SEPM::evaluate()} as argument \code{comparison}.
}
\details{
Specification of correlation structure: always starts with "type=..." specifying the
correlation matrix type. Currently, only "type=equi" (equicorrelation) and
"type=ak" (autocorrelation) are supported.
Equicorrelation may be specified as "type=equi_rho=x" (e.g. "type=equi_rho=0.5") for correlation
matrix with entries R[i,j]=x^(|i-j|>0)
Autocorrelation may be specified as "type=ak_rho=x" (e.g. "type=ak_rho=0.5") for correlation
matrix with entries R[i,j]=x^|i-j|.
}
\examples{
dat <- sample_lfc_acc()
str(dat, 2)
dat$comp

}
