% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mbb.R
\name{sample_mbb_acc}
\alias{sample_mbb_acc}
\title{Sample data from multivariate beta-binomial model}
\usage{
sample_mbb_acc(
  gd = "nu40_rho50_blocks2",
  params = NULL,
  n = 200,
  data = NULL,
  job = NULL,
  nrep = 1,
  ...
)
}
\arguments{
\item{gd}{character, generative distribution, needs to be one of \code{names(DIST)}}

\item{params}{alternatively to \code{gd}, params can also be specified directly,
e.g. via \code{params = SIMPle::define_dist()[[1]]$params}}

\item{n}{integer, total sample size}

\item{data}{ignored (required for batchtools compatibility)}

\item{job}{ignored (required for batchtools compatibility)}

\item{nrep}{integer, if != 1, then output is list with nrep repetitions}

\item{...}{further arguments (ignored)}
}
\description{
Sample data from multivariate beta-binomial model
}
