% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_mbb.R
\name{sample_pars}
\alias{sample_pars}
\title{Sample parameters from multivariate binomial distribution}
\usage{
sample_pars(nu = NULL, moments = NULL, gamma = NULL, ...)
}
\arguments{
\item{nu}{numeric, concentration parameter}

\item{moments}{matrix, moment matrix}

\item{gamma}{numeric, concentration vector}

\item{...}{further arguments (ignored)}
}
\description{
Sample parameters from multivariate binomial distribution
}
