% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_mle_core.R
\name{save_results_mle}
\alias{save_results_mle}
\title{Save MLE simulation results in tabular form}
\usage{
save_results_mle(reg, JP.data, target, return = FALSE)
}
\arguments{
\item{reg}{batchtools registry}

\item{JP.data}{data.table, jobPars of MLE data generation registry}

\item{target}{character, target directory}

\item{return}{logical, should the object also be returned?}
}
\description{
Save MLE simulation results in tabular form
}
