% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_mle_study.R
\name{select_setM}
\alias{select_setM}
\title{Select a subset of models}
\usage{
select_setM(
  instance,
  methods = NA,
  M = 200,
  M.start = NA,
  M.probs = c("uniform", "learn.theta", "learn.theta.squared"),
  M.seed = 1
)
}
\arguments{
\item{instance}{instance generated by \code{sample_mle}}

\item{methods}{character, allowed methods e.g. "glmnet|xgbTree", if NA all methods are allowed
(i.e. NA is a shortcut to "glmnet|xgbTree|rpartCost|svmLinearWeights2")}

\item{M}{integer, number of models}

\item{M.start}{integer, first model; if M.start=NA, models are choosen randomly according to M.probs}

\item{M.probs}{character, one of "uniform", "learn.theta", "learn.theta.squared"}

\item{M.seed}{integer, random seed for random draw of models}
}
\description{
Select a subset of models
}
