% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_lfc.R
\name{study_lfc_acc}
\alias{study_lfc_acc}
\title{Conduct an evaluation study (classification accuracy)}
\usage{
study_lfc_acc(
  instance = sample_lfc_acc(),
  alternative = "greater",
  alpha = 0.025,
  estimate = "beta.approx",
  infer = "maxT",
  transform = "none",
  data = NULL,
  job = NULL
)
}
\arguments{
\item{instance}{artificial performance data, result of \code{\link{sample_lfc_acc}(...)} call}

\item{alternative}{character, either \code{"two.sided"} (default), \code{"greater"} or \code{"less"}}

\item{alpha}{numeric, significance level (default: \code{0.05})}

\item{estimate}{character, estimation method (default: \code{"default"}, i.e. ML estimation)}

\item{infer}{character, inference method (default: "maxT", currently only choice)}

\item{transform}{character, application of transform and delta method (default: \code{"none"},
other suitable choices are \code{"logit"} and \code{"asin.sqrt"})}

\item{data}{ignored (required for batchtools compatibility)}

\item{job}{ignored (required for batchtools compatibility)}
}
\value{
A single row \code{data.frame} with important characteristics of the evaluation study.
}
\description{
This is a \code{batchtools} friendly implementation of the evaluation of the artificial
prediction models simulated via
\code{\link{sample_lfc_acc}} regarding classification accuracy with functions from the
\code{SEPM} package.
}
\examples{
study_lfc_acc(instance=sample_lfc_acc())

}
