% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_mbb.R
\name{study_mbb_acc}
\alias{study_mbb_acc}
\title{Analyze synthetic data from multivariate beta-binomial model.}
\usage{
study_mbb_acc(
  instance = sample_mbb_acc(),
  prob = 0.95,
  method = c("sample", "copula", "approx"),
  prior = c("vague", "correct", "liberal"),
  contrast = c("raw", "rdm"),
  data = NULL,
  job = NULL
)
}
\arguments{
\item{instance}{result of \code{sample_mbb_acc()}}

\item{prob}{numeric, coverage probability}

\item{method}{character, inference method ("sample", "copula" or "approx")}

\item{prior}{character, specifies the prior assumptions ("vague", "correct" or "liberal")}

\item{contrast}{character, specifies the contrast to be applied ("raw" or "rdm")}

\item{data}{ignored (required for batchtools compatibility)}

\item{job}{ignored (required for batchtools compatibility)}
}
\description{
Analyze synthetic data from multivariate beta-binomial model.
}
