% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_mle_core.R
\name{summarize_results}
\alias{summarize_results}
\title{Summarize simulation study in results table}
\usage{
summarize_results(
  data,
  vars = "final.theta",
  rules = c("oracle", "default", "within1SE", "optimalEFP"),
  comp = c(4, 3),
  NE = list(400, 800),
  NL = list(c(400, 800)),
  alpha = 0.01,
  digits = 3
)
}
\arguments{
\item{data}{data.frame, containing simulation results}

\item{vars}{character, list of variables to report}

\item{rules}{character vector, selection rules to investigate}

\item{comp}{numeric of length 2, which two selection rules should be compared (delta = rules[comp[1]] - rules[comp[2]])}

\item{NE}{list, each element is a numeric vector hinting which evaluation sample sizes (n.eval) are considered}

\item{NL}{list, each element is a numeric vector hinting which learning sample sizes (n.learn) are considered}

\item{alpha}{numeric, significance level of t-test in comparison}

\item{digits}{numeric, to which degree should results be rounded}
}
\value{
data.frame with aggregated results
}
\description{
Summarize simulation study in results table
}
