% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_mle_core.R
\name{save_instance}
\alias{save_instance}
\title{Save MLE simulations instances to disk}
\usage{
save_instance(
  instance,
  id = instance$job$id,
  filename = "mle_instance_",
  folder = "E:\\\\MLE_SIM\\\\DATA",
  sep = "\\\\",
  write = TRUE,
  data = NULL,
  job = NULL
)
}
\arguments{
\item{instance}{object generated via \code{sample_mle}}

\item{id}{integer, id to append to \code{filename}}

\item{filename}{character, base file name}

\item{folder}{character, folder to file}

\item{sep}{character, path separator}

\item{write}{logical, should instance be saved a (if FALSE, instance is only returned)}

\item{data}{NULL (required for batchtools compatibility)}

\item{job}{NULL (required for batchtools compatibility)}
}
\description{
Save MLE simulations instances to disk
}
